refactor(architecture): move usecases to domain modules
Consolidate usecases into respective domain modules following DDD principles.

Eliminate UsecasesModule and resolve circular dependencies.
