+++ src/briefings/usecases/dto/categorize-articles.dto.ts
