+++ src/briefings/usecases/dto/generate-brief.dto.ts
