+++ src/briefings/usecases/dto/process-articles.dto.ts
