+++ src/briefings/usecases/dto/rate-articles.dto.ts
