+++ src/briefings/usecases/dto/scrape-articles.dto.ts
