+++ src/briefings/usecases/generate-brief.usecase.ts
