+++ src/briefings/usecases/generate-simple-brief.usecase.ts
