--- src/usecases/usecases.module.ts
@@ -1,70 +1,0 @@
-import { S3Module } from '@libs/s3';
-import { Module, forwardRef } from '@nestjs/common';
-import { AiModule } from '../ai/ai.module';
-import { AudioFilesModule } from '../audio-files/audio-files.module';
-import { BriefingModule } from '../briefing/briefing.module';
-import { ConfigModule } from '../config/config.module';
-import { ProcessorModule } from '../processor/processor.module';
-import { ProfilesModule } from '../profiles/profiles.module';
-import { ScraperModule } from '../scraper/scraper.module';
-import { YoutubeChannelsModule } from '../youtube-channels/youtube-channels.module';
-import { YoutubeTranscriptionsModule } from '../youtube-transcriptions/youtube-transcriptions.module';
-import { GenerateAudioUseCase } from './audio/generate-audio.usecase';
-import { CategorizeArticlesUseCase } from './briefing/categorize-articles.usecase';
-import { GenerateBriefUseCase } from './briefing/generate-brief.usecase';
-import { GenerateSimpleBriefUseCase } from './briefing/generate-simple-brief.usecase';
-import { ProcessArticlesUseCase } from './briefing/process-articles.usecase';
-import { RateArticlesUseCase } from './briefing/rate-articles.usecase';
-import { RunBriefingUseCase } from './briefing/run-briefing.usecase';
-import { ScrapeArticlesUseCase } from './briefing/scrape-articles.usecase';
-import { ExtractYoutubeTranscriptsUseCase } from './youtube-transcriptions/extract-youtube-transcripts.usecase';
-import { ListTranscriptionsUseCase } from './youtube-transcriptions/list-transcriptions.usecase';
-import { ProcessTranscriptionFilesUseCase } from './youtube-transcriptions/process-transcription-files.usecase';
-
-@Module({
-  imports: [
-    ScraperModule,
-    forwardRef(() => ProcessorModule),
-    BriefingModule,
-    ProfilesModule,
-    YoutubeChannelsModule,
-    forwardRef(() => YoutubeTranscriptionsModule),
-    ConfigModule,
-    AiModule,
-    S3Module,
-    AudioFilesModule,
-  ],
-  providers: [
-    // Audio usecases
-    GenerateAudioUseCase,
-    // Briefing usecases
-    ScrapeArticlesUseCase,
-    ProcessArticlesUseCase,
-    RateArticlesUseCase,
-    CategorizeArticlesUseCase,
-    GenerateBriefUseCase,
-    GenerateSimpleBriefUseCase,
-    RunBriefingUseCase,
-    // YouTube transcription usecases
-    ExtractYoutubeTranscriptsUseCase,
-    ProcessTranscriptionFilesUseCase,
-    ListTranscriptionsUseCase,
-  ],
-  exports: [
-    // Audio usecases
-    GenerateAudioUseCase,
-    // Briefing usecases
-    ScrapeArticlesUseCase,
-    ProcessArticlesUseCase,
-    RateArticlesUseCase,
-    CategorizeArticlesUseCase,
-    GenerateBriefUseCase,
-    GenerateSimpleBriefUseCase,
-    RunBriefingUseCase,
-    // YouTube transcription usecases
-    ExtractYoutubeTranscriptsUseCase,
-    ProcessTranscriptionFilesUseCase,
-    ListTranscriptionsUseCase,
-  ],
-})
-export class UsecasesModule { }
