+++ src/youtube-transcriptions/usecases/dto/extract-youtube-transcripts.dto.ts
