+++ src/youtube-transcriptions/usecases/dto/process-transcription-files.dto.ts
