+++ src/youtube-transcriptions/usecases/process-transcription-files.usecase.ts
